+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Make controller Docker image `tee` to the logfile, rather than redirecting all output
 * Implement a 10-second HTTP request timeout in the Gecko client
 * Add comments and documentation in RPC workflow and chit spammer tests
+* Rename HighLevelGeckoClient to RpcWorkflowRunner and move to ava_testsuite package
 
 # 0.6.0
 * Use Kurtosis version that allows the user to configure network width